Server refactoring + Logger (#13)
This commit contains new Webserver features:

1. Access and Error loggers. Error logger should be used more often in different parts of Server code, I'll implement this later.

The Access logger outputs the log in a Extended Log Format
2. The server class has been completely redesigned and a new architecture has been applied. The server is now more extensible.

Bug fixes(Only one :<):

Fixed a case where the server would silently terminate when nginx was running on the same port as the web server. The web server now displays the reason for termination.